Add VHLO Python/C Bindings (#1142)
These are pretty minimal VHLO bindings. We don't intend for Python/C
users to directly interact with VHLO attributes / types, so we only need
the ability for Python to register the dialect to parse and serialize.

Currently this is based on CHLO bindings, except only Dialect-related
bindings are provided currently, no attributes or types.