Organize ongoing work items in workstreams and milestones
As we've been exploring the StableHLO opset, we've been making notes
about all kinds of todo items - starting from minor cleanups in the
StableHLO dialect all the way to missing functionality in the opset.

This is great because thanks to that we have documented at great length
what it would take for StableHLO to reach its full potential - with
perfect frontend contract, 100% specification compliance, etc etc.

This level of detail (at the moment of writing, we have 275 tickets in
our issue tracker) would benefit from some organization, and this is
something that this PR proposes.

Recently, we have categorized the tickets in the issue tracker into
multiple workstreams and tied these workstreams to the milestones.
The roadmap.md introduced in this PR brings this all together.