Prototype a mechanism to customize shape refinement for CustomCallOp
This is a prototype of the CustomCallOp shape refinement mechanism to
enable specializing dynamically-shaped programs with custom calls to
static shapes (#851).

Based on the results of prototyping, I'm planning to propose an RFC
to extend CustomCallOp, so that the logic can migrate from using an
unregistered attribute to something more solid.