Move and merge verifier of Set/GetDimensionSizeOp (#1029)
SetDimensionSizeOp is very special, compared to other ops when involving
bounded dynamism: it is possible that the output have bounds while the
inputs has no bounds at all. Thus we can't use this util function to
generate the encoding at the end of shape function:
```
Attribute boundsToEncoding(Attribute prototype, ArrayRef<int64_t> bounds)
```
Because the `prototype`, which comes from `input.getEncoding()` maybe
null. So in this PR it create bounds attr from dialect directly.

Change `inferReturnTypes` to `inferReturnTypeComponents` and remove the
TODO in comment as well.