Replace VHLO_UnitAttr with VHLO_BooleanAttr
This is one of the several steps towards aligning VhloAttrs.td with
the StableHLO spec.

As far as representational power goes, UnitAttr is equivalent
to BooleanAttr, so in the spirit of VHLO being a shallow dialect, I
think that we should remove the former and use the latter instead.

When making this change, I noticed that BooleanAttr attributes from
StableHLO get translated to IntegerAttr attributes from VHLO. This has
been changed to use the newly introduced BooleanAttr from VHLO.

This is a breaking change for VHLO, and I was hesitating whether to
provide proper versioning for it or not. Ultimately, I decided that
I won't be doing that, similarly to how the work on forking attributes
and types (also a breaking change) didn't do that. We haven't yet
officially released VHLO, so I think that it's fine.