Inline VHLO_ConvDimensionNumbersAttr and friends
I think that grouping of the elementary attrs into bigger attrs is
pretty incidental as far as serialization format goes, so I would
propose to break all of them down, just like the spec does.

This is a breaking change for VHLO, and I was hesitating whether to
provide proper versioning for it or not. Ultimately, I decided that
I won't be doing that, similarly to how the work on forking attributes
and types (also a breaking change) didn't do that. We haven't yet
officially released VHLO, so I think that it's fine.