Remove all references to 0.3.0 and 0.4.0
Having two distinct unofficial versions of StableHLO was useful when
developing the versioning machinery, but once we release 0.9.0, this
will likely be confusing because we never announced any other versions
(and rightfully so - they were not supposed to be public).

In preparation for v0.9.0, let's remove all references to 0.3.0 and
0.4.0, squash all versions of VHLO ops and flush the VHLO versioning
patterns.

Closes #1203.