Refactor StableHLO <=> VHLO
In the last few pull requests, I crammed a lot of adhoc logic into
the corresponding conversion patterns.

When working on removing optional attrs from VHLO, I needed to further
update the conversion patterns, and I found it to be unnecessarily
complicated with all the stuff in there.

This PR refactors the conversion patterns to encapsulate the logic
that bridges the gaps between StableHLO and VHLO. I think that modifying
the conversion patterns will become considerably easier now.