Use _v1 in VHLO mnemonics (#1235)
We've been oscillating on vhlo.foo vs vhlo.foo_v1 in the last few
months, but vhlo.foo_v1 has won because of its explicitness - we have
encountered quite a few questions along the lines of "if I see vhlo.foo,
what version does it have?".

In the current design, VHLO isn't supposed to be manipulated by users
directly, so we figured that the downside of increased verbosity is
going to be relatively minor in practice. If this ends up not being the
case, then this decision isn't going to be set in stone - we will always
be able to drop _v1 before the 1.0 release.