Only lint whitespace for changed files (#1252)
Now that we have a bunch of test files, it saves a lot of time to only
lint changed files.

Granted we will likely be shedding most of these test files soon
(reorganizing, trimming, changing where they are stored, or something
else), this is still a worthwhile change to make.

Also updated the fix link so that it is copy-pastable. This tool is only
runnable from the root directory as it is, so including full relative
path is probably best.