Improve newcomer experience
I've tried to follow our build instructions on my personal macbook to
see how smooth the experience is and how onerous it is to run the
testdata/ suite on a reasonably mainstream machine.

Turns out that our instructions don't really work on macOS, and
predictably checking the entire testdata/ takes a considerable amount
of time (20+ seconds on a 2021 M1 Pro macbook).

This PR fixes the build instructions to work on macOS and updates the
CMake target called out in the readme to only check the tests/ suite.