Add interpreter for PowOp (#1207)
closes #981 

We have the following constraints in the spec:

```
(I1) lhs is a tensor of integer, floating-point, or complex type.
(I2) rhs is a tensor of integer, floating-point, or complex type.
(C1) lhs, rhs, and result have the same type.
```

These constraints are covered by the following tests

```
(I1) lhs is not a tensor of integer, floating-point, or complex type. (covered by ODS)
(I2) rhs is not a tensor of integer, floating-point, or complex type. (covered by ODS)
(C1) lhs, rhs, and result do not have the same type. (covered by ODS)
```

Add a binary version of mapWithUpcastToDouble function.