Link recently added frontend contract tickets in the spec (#1299)
Today I realized that all "Consider removing" tickets are actually
Frontend Contract tickets, not New Features. If we don't clean them up
now in preparation for StableHLO v1.0, we will have to support them for
a long long time.

The outcome of some of these tickets might actually be "we considered
removing this feature because it has problems, but it's actually in use
so we'd need to introduce a new feature to replace it", but we still
need to get them through the Frontend Contract phase.

This PR updates the spec to call out these tickets. Also, it cleans up
mentions of Specification Compliance tickets because those belong in the
StableHLO dialect (there can be multiple implementations of the spec, so
we shouldn't be preferring one of them in the spec by linking to its
tickets).