Streamline the on: part of all GitHub workflows (#1310)
I was looking to fix the long-standing issue with checks being skipped
by paths/paths-ignore, leading to: 1) spuriously blocked merges if the
checks are required, 2) weird "in progress" dots next to PRs.

I discovered that there is no easy solution to this, but there is a
workaround:
https://docs.github.com/en/repositories/configuring-branches-and-merges-in-your-repository/defining-the-mergeability-of-pull-requests/troubleshooting-required-status-checks#handling-skipped-but-required-checks.

This solution works, but it's quite fiddly. Defining a shadow workflow,
making sure that its paths/paths-ignore are an exact negation of the
real workflow - this got old pretty fast.

Then I realized that we don't have to suffer through this. We can just
enable all our workflows on all pull requests. Now that we have fast CI
runners, waiting for unnecessary builds when we only update docs would
take a few minutes at most, and I think it's a worthy tradeoff for
getting rid of awkwardness.

So I went and removed all paths/path-ignore from all our workflows. When
doing that, I noticed that their on: parts are different between one
another, so I went ahead and unified that too. If we're running builds
on pushes, we might as well run lints on pushes too.