Fix tests by updating STABLEHLO_TOOLS_DIR
`set(STABLEHLO_TOOLS_DIR ${CMAKE_BINARY_DIR}/bin)` is the right way
of obtaining a path to the directory that contains stablehlo-opt and
stablehlo-translate.

CMAKE_CURRENT_BINARY_DIR works as well if we're building standalone,
but that breaks down when building as part of another project, which
we discovered when integrating into MLIR-HLO. This pull request fixes
this oversight.