refactor to improve evalOp function type checking (#1347)
This is to make sure that, during compilation, the
interfatce-type-checking for the `evalOp*` function calls in eval
function
(https://github.com/openxla/stablehlo/blob/311f14ce78c7fe35d304ee91007b58c335cf821e/stablehlo/reference/Ops.cpp#L484)
are relying on the declarations in Ops.h. Currently, the checking is
happening locally within Ops.cpp as the calls to those `evalOps*`
functions are preceded by the corresponding definitions. The problem
with the current setup is that a change in declaration Ops.h will go
unnoticed, which is what exactly happened in
https://github.com/openxla/stablehlo/pull/1343. The proposed fix will
make sure that such inconsistencies can be caught during compilation
time.