Drop indices_of_shape_operands once they are no longer needed
--stablehlo-canonicalize-dynamism can now drop indices_of_shape_operands
once they are no longer needed, similarly to how it transforms
dynamic_foo ops to foo ops when operands like output_shape are no
longer needed.

To make sure that we only do the transformation on valid IR, I also
wrote verification logic for indices_of_shape_operands as well as tests
for this logic.

In the future, when we upgrade indices_of_shape_operands from an
experiment to a full-fledged StableHLO feature, this logic will be moved
to a proper verifier.