Harden --stablehlo-canonicalize-dynamism
When working on improvements for indices_of_shape_operands, I noticed
that --stablehlo-canonicalize-dynamism canonicalized some operations
that it shouldn't have.

I figured that I should add negative tests for all failure cases in
StablehloCanonicalizeDynamismPass, and this uncovered several issues
in the logic of the pass:
  * DynamicBroadcastInDimOp::output_dimensions,
    DynamicIotaOp::output_shape, DynamicReshapeOp::output_shape were
    discarded even if they weren't constants.
  * DynamicGatherOp/DynamicPadOp/RealDynamicSliceOp canonicalizations
    were using type inference to create static versions of the ops,
    which means that they could change the original result type,
    potentially producing invalid programs if someone downstream was
    relying on this exact result type.

This pull request provides fixes to those issues and comprehensively
covers all notifyMatchFailure cases in
StablehloCanonicalizeDynamismPass to make sure that this doesn't happen
again.