Extend indices_of_shape_operands to work for tuple types
NOTE: This pull request is based on #1386.

At the moment, indices_of_shape_operands are in 1:1 correspondence with
result types, but that doesn't work when custom calls return tuples.

This pull requests addresses this problem, and changes
indices_of_shape_operands to be in 1:1 correspondence with flattened
result types. Originally, I thought that I'd need to redesign
indices_of_shape_operands, e.g. carry arrays of 1-dimensional tensors or
arrays of strings, but this solution keeps the existing structure of
the attribute, i.e. a 1-dimensional tensor of i64.

The changes are surprisingly compact and cover the following three
areas of the implementation:
  1) Verification logic for the attribute that lives in
     getShapeRefinements in Base.h.
  2) Helper logic which applies refinements to operation types
     that lives in refineReturnTypes in StablehloRefineShapes.cpp.
  3) inferMostSpecificType logic which is used to merged unrefined
     types and the corresponding refinements.

These changes also led to an unexpected improvement. Now that
--stablehlo-refine-shapes started operating on tuple types,
tensor::CastOp stopped working (it only applies to tensors), so I was
forced to look for a better solution and remembered about
UnrealizedConversionCastOp. This was an easy migration, and as a nice
bonus the pass no longer depends on the Tensor dialect.