Add interpreter for SignOp. (#1382)
closes #1115 

We have the following constraints in the spec:

```
(I1) operand is a tensor of signed integer, floating-point, or complex type.
(C1) operand and result have the same type.
```

These constraints are covered by the following tests

```
I1: a) operand is not a tensor of signed integer, floating-point, or complex type. (Covered by ODS).
C1: a) type(operand) != type(result). (Covered by ODS).
```