Add interpreter for GetDimensionSizeOp (#1436)
Here are the constraints for the GetDimensionSizeOp:
```
(I1) operand is a tensor.
(I2) dimension is a constant of type `si64`.
(C1) 0 <= dimension < rank(operand).
```

These constraints will be comprehensively covered by the following
tests:
```
I1: a) operand is not a tensor. (Covered by ODS).
I2: a) dimension is not a constant of type `si64`. (Covered by ODS).
C1: a) 0 < dimension.
    b) dimension < rank(operand).
```

If we drop the "Covered by ODS" pieces, this will leave us with the
following test cases:
```
C1a: 0 < dimension.
C1b: dimension < rank(operand).
```

closes #1432