Assorted refactorings of the interpreter
Here are some of the things that I noticed over the last few weeks
but didn't have the time to follow up on:

1) Let's unify the boilerplate in the eval() function to: a)
consistently use auto, b) drop the "runtime" part of variable names to
make things easier to read, c) consistently declare temporary variables.
As an alternative to c), we could skip creating temporaries except
for result.

2) Let's drop isSupportedFooType checks in Element.cpp when there's
just one supported category of types. These checks are redundant
because getFooValue will check that anyway.

3) Same for the isSupportedComplexType + comparisonDirection check
in evalCompareOp. It is redundant as well.

4) There are few more minor cleanups in this pull request which I don't
think need special description / justification.