Add interpreter for CbrtOp (#1417)
Here are the constraints for CbrtOp:
```
(I1) operand is a tensor of floating-point or complex type.
(C1) `operand` and `result` have the same type.
```
I1 and C1 are covered by the ODS, so no additional tests are added.

Notes:
* The implementation is inspired from De Moivre's formula to calculate
nth root of a complex number. k is assumed 0 for principal root. See:
https://en.wikipedia.org/wiki/De_Moivre%27s_formula

closes #1099