Refactor EvalSliceOpPattern to use APSInt-based matchInts (#1452)
Erasing the bitness and the signedness of the underlying values to
int64_t is reasonable when we're going to use these values in the
int64_t context (e.g. take a shape of DynamicReshapeOp and then put it
into ShapedTypeComponents).

However, in the case of data movement ops like BroadcastInDimOp,
ConcatenateOp and now SliceOp, this doesn't make much sense.

This even caused a crash when matchInts was called when trying to
evaluate slices of unsigned tensors. When we were trying to call
DenseIntElementsAttr::get with an unsigned tensor type and values of
type SmallVector<int64_t>, that led to a crash.

Now that we have switched to SmallVector<APSInt>, this crash is fixed.
I've also audited all occurrences of DenseIntElementsAttr::get to make
sure that we no longer have any type mismatches that could lead to
further crashes elsewhere.