Specification for quantized  AddOp (#1446)
## Summary 

The PR proposes the specification for quantized add op.

## A few details

At some point we
[decided](https://github.com/openxla/stablehlo/pull/1352#discussion_r1166196224)
to drop the introduction of the specification of this op mainly because
we were unsure about the fate of
https://github.com/openxla/stablehlo/issues/1406.
 
Please have a look at my revised proposal on
https://github.com/openxla/stablehlo/issues/1406 and let me know if I am
missing something. Otherwise, let us review this op and let me know your
feedback.

Side note: For those who are already aware of the context of prior
introduction of this op, please note that the current proposal is almost
same as before except that it does not have any additional constraint
imposed by the op's semantics on `storage_min` or `storage_max`.