Specification for quantized DotGeneralOp (#1413)
## Summary 
The PR proposes the spec for quantized dot-general op along with the
specifications for a few other ops on which the dot-general depends on,
for example, `slice`, `transpose`, and `reshape`.

## A few details
Given `fp = tensor with floating-point type and q = tensor with
uniformed quantized type`, the PR covers the semantics of
(1) Static range quantized `dot_general` op `dot_general(q, q)`, and 
~~(2) Hybrid quantized `dot_general` op `dot_general(fp, q)`: Currently,
this version of the op only supports dynamic range quantization, where
the on-the-fly quantization of `lhs` is fused in the op-semantics. IMO,
once we support https://github.com/openxla/stablehlo/issues/1407, the
quantization logic can be un-fused and made explicit in the MLIR graph
(cc @sngyhan).~~

**update**: As per the
[discussion](https://github.com/openxla/stablehlo/pull/1413#discussion_r1183127043),
it is decided to have only (1) in the spec. It might be too early to
introduce (2), the "dynamic range quantizated" variant of the op, mainly
because (a) only TFLite CPU implements it and (b) in the long, there are
plans to implement dynamic range quantization expolicitly in the graph
level.


## What comes next
The plan forward is to propose a PR for convolution op in very near
future. I realized that the spec for convolution depends on dot-general
and a split might help the review process.

Please let me know your review feedback.