Add interpreter for CholeskyOp (#1444)
Here are the constraints for CholeskyOp:
```
(I1) `a` is a tensor of floating-point or complex type.
(I2) `lower` is a tensor constant of `i1` type.
(C1) `a` and `result` have the same type.
(C2) rank(`a`) >= 2.
(C3) dim(`a`, -2) = dim(`a`, -1).
```

These constraints will be comprehensively covered by the following
tests:

```
I1: a) `a` is not a tensor of floating-point type or complex type. (Covered by ODS).
I2: a) `lower` is not a tensor constant of `i1` type. (Covered by ODS).
C1: a) type(a) != type(result).
C2: a) rank(a) < 2.
C3: a) dim(`a`, -2) != dim(`a`, -1).
```

If we drop the "Covered by ODS" pieces, this will leave us with the
following test cases:

```
C1a: type(a) != type(result).
C2a: rank(a) < 2.
C3a: dim(`a`, -2) != dim(`a`, -1).
```

Notes:
* Implementation inspired by the [Cholesky–Banachiewicz
algorithm](https://en.wikipedia.org/wiki/Cholesky_decomposition).

closes #1123