Integrate tosa/ into the Bazel build
This PR merges the BUILD.bazel file from stablehlo/conversions/tosa/transforms
into the root BUILD.bazel file, to follow the current convention.

We cargo-culted this convention from the MLIR-HLO repository, so maybe
it's time to get rid of it, but I'll leave that to future work.

Furthermore, I added BUILD.bazel in stablehlo/conversions/tosa/tests,
so that Bazel can actually run those tests.

Finally, I noticed that a minor cleanup opportunity in the CMake build,
and I figured it wouldn't hurt to pursue it in this PR.