Rename tosa/ passes to follow pass name conventions (#1504)
For the legalization passes, we've been following the
foo-legalize-to-bar convention. And for passes in general, we've been
adhering to dialectname-something-something. These conventions are
applied to the recently introduced tosa/ passes in this PR.

Also, to further the consistency between existing passes and newly
introduced passes, this PR removes manual definitions of createFooPass
functions. At some point, TableGen got the ability to automatically
generate these definitions, so we're leveraging it here.

Thank you, @GleasonK for your feedback that led to the creation of this
pull request.