Actually use the BytecodeWriterConfig in writeBytecodeToFile (#1512)
#1511 introduced BytecodeWriterConfig to our serialization logic,
but I forgot to actually use it. This PR fixes this oversight.

This mistake on my part happened because we don't have forward
compatibility tests yet. To make sure that I didn't mess up anything
else this time, I have manually verified the following:
* At HEAD, build/bin/stablehlo-translate --serialize produces different
payloads when: 1) local clone of llvm-project is pristine, 2) local
clone of llvm-project has kVersion manually changed to 2. The only
difference is the bytecodeVersion field in the serialized payloads.
* With this PR, build/bin/stablehlo-translate --serialize produces the
same payloads when: 1) local clone of llvm-project is pristine, 2) local
clone of llvm-project has kVersion manually changed to 2. As expected,
using BytecodeWriterConfig overrides kVersion.