Delete testdata bytecode tests (#1425)
Recently we gathered some code coverage metrics for testdata tests, and
compared that to our unit tests.

It turns out that our testdata coverage was a strict subset of unit
tests, including in generated files, which makes them not particularly
useful as serialization and deserialization tests, and minimally useful
as versioned semantic compatibility tests, but very useful as testdata
for StableHLO users.

For semantic compatibility, this approach does not scale well due to the
high redundancy in test cases. The overhead for cloning all these files
at each StableHLO version bump is too high for its limited value (25MB
of bytecode files per version, difficult to review due to PR size,
consistently breaks our tooling). We have created #1416 to rethink
semantic compatibility testing, and are planning to only version
`stablehlo_legalize_to_vhlo.mlir` for serialization testing, which alone
has 95% coverage of compatibility machinery, covering all ops, types,
and attributes.


This PR was created using the following commands (meaning it should be
uniform):

```
cd stablehlo/testdata
rm *.mlir.bc
sed -i 's/stablehlo-translate.*--interpret/stablehlo-opt -inline %s | stablehlo-translate --interpret/' *.mlir
sed -i "/0_9_0/d" *.mlir
```

**EDIT 5/24: Alternatives considered and mitigations**

A primary use case for testdata on top of reference interpreter testing,
is to provide copy-pastable snippets for StableHLO users to test their
backend implementations against. As such, maintaining some form of
textual assembly format is a requirement. The primary alternative
considered was to remove the MLIR files and only have bytecode files.
This is more stable, and there is an MLIR vscode extension that permits
viewing and editing bytecode files in place. This solution was not
chosen for two main reasons - 1. Serialized portable artifacts are VHLO,
meaning the decoded file would not be a copy-pastable snippet, as IR
upgrades and conversion to StableHLO are still required. 2. This
extension does not work on github, meaning files would need to be opened
in VSCode or deserialized using an opt tool.

The downside of only preserving text files is that StableHLO portable
artifacts have stability, meaning they will not break between releases.
Textual assembly format may break. In the case of assembly format
changes, we may have a large amount of testdata that breaks and requires
fixing. This overhead can be automated away with a script that checks
out a known good release, generates bytecode files, and deserializes at
HEAD to update the testdata. This script is tracked in #1533.