Make sure that getBytecodeVersion is compatible with upstream
Before calling writeBytecodeToFile, we also need to ensure that the
bytecode version that we're requesting falls within the range of
supported bytecode versions of the MLIR version that we're using.

Surprisingly, this isn't done by MLIR (e.g. the current behavior for
setDesiredBytecodeVersion is to clamp the provided bytecode version
to the range of supported versions, which may result in payloads
whose StableHLO version doesn't match the expected bytecode version).