Align the notation of spec and interpreter for Data Movement operations
Data Movement operations were one of the first operations specced,
so we were still working out the notation back then.

As mentioned in #484, inspired by our experience with the interpreter,
it appears to be useful to express semantics of these operations in
terms of result_index / operand_index rather than in terms of individual
components of those indices.

I took a stab at tweaking the specification for these operations, and
I think that going from scalars to tensors to represent indices results
in more readable specification. Furthermore, descriptive names like
result_index are more conducive to readability than i or j. Also,
we can retire the somewhat weird "id" / "jd" notation for these ops.

Overall, I think that the specification and the implementation of
data movement ops are well-aligned with each other. A few quirks:
  * Both PadOp and TransposeOp are formulated in terms of operand's
    iteration space because they are much easier to implement this way.
    Everything else is formulated in terms of result's iteration space.
  * SortOp doesn't have a low-level 1:1 mapping, but there's a 1:1
    mapping of high-level pieces and comments help with understanding.

Re: the rest of the opset:
  * Control Flow ops look good as well (modulo a minor simplification
    in #1553).
  * Data Movement ops are discussed in this PR.
  * Distribution ops are in the works and will be reviewed afterwards.
  * Dynamism ops are in the works as well, and we'll first need to
    spec them before worrying about the interpreter. Overall, I'm not
    worried because at runtime there's not going to be a difference
    between static ops and their dynamic counterparts.
  * Elementwise ops are straightforward as far as notation goes.
  * Extensibility ops are out of scope for now because we may decide
    to remove tuples altogether (#598).
  * Most of the Miscellaneous ops are going to be handled via passes
    in the long run, so I haven't reviewed those. The rest, i.e.
    constant and iota, are fine with a minor fix to iota.
  * Not in HLO ops are going to be out of the opset in the long run,
    so we haven't specced them and aren't planning to spec or implement
    them.
  * Quantization ops are elementwise, so they are straightforward
    too as far as notation goes.
  * Finally, Reduction ops are going to be tricky because their specs
    have been tricky, their implementations have been tricky, and their
    notations have been drifting apart. I opened #1551 to revisit this
    once we're done with the first round of implementations.